refactor: split pi-devtools into maintainable modules
